+++ src/layer-error.js
@@ -169,6 +169,7 @@ LayerError.dictionary = {
   predicateNotSupported: 'The predicate is not supported for this value of model',
   invalidPredicate: 'The predicate does not match the expected format',
   appIdImmutable: 'The appId property cannot be changed',
+  clientMustBeReady: 'The Client must have triggered its "ready" event before you can call this',
 };
 
 module.exports = LayerError;